Use -rpath to support shared objects properly.

+++ ext/extconf.rb
@@ -5,7 +5,11 @@ HERE = File.expand_path(File.dirname(__FILE__))
 BUNDLE = Dir.glob("libmemcached-*.tar.gz").first
 BUNDLE_PATH = BUNDLE.sub(".tar.gz", "")
 
-DARWIN = `uname -sp` == "Darwin i386\n"
+TARGET = if `uname -sp` == "Darwin i386\n"
+  :darwin
+else
+  :linux
+end
 
 if !ENV["EXTERNAL_LIB"]
   $includes = " -I#{HERE}/include"
@@ -13,8 +17,9 @@ if !ENV["EXTERNAL_LIB"]
 
   $CFLAGS = "#{$includes} #{$libraries} #{$CFLAGS}"
   $LDFLAGS = "#{$libraries} #{$LDFLAGS}"
-  $CPPFLAGS = $ARCH_FLAG = $DLDFLAGS = ""
+  $LIBS << " -lmemcached"
 
+  $CPPFLAGS = $ARCH_FLAG = $DLDFLAGS = ""
   $LIBPATH = ["#{HERE}/lib"]
   $DEFLIBPATH = []
 
@@ -25,28 +30,27 @@ if !ENV["EXTERNAL_LIB"]
       puts "Building libmemcached."
       puts(cmd = "tar xzf #{BUNDLE} 2>&1")
       raise "'#{cmd}' failed" unless system(cmd)
+      
+      env = case TARGET
+        when :solaris then "CFLAGS='-fPIC -R=#{HERE}/lib' LDFLAGS='-fPIC -rpath=#{HERE}/lib'"
+        when :darwin, :linux then "CFLAGS='-fPIC -rpath=#{HERE}/lib' LDFLAGS='-fPIC -rpath=#{HERE}/lib'"
+      end
 
       Dir.chdir(BUNDLE_PATH) do
-        puts(cmd = "env CFLAGS=-fPIC LDFLAGS=-fPIC ./configure --prefix=#{HERE} --without-memcached --disable-shared --disable-utils 2>&1")
+        puts(cmd = "env #{env} ./configure --prefix=#{HERE} --without-memcached --disable-utils 2>&1")
         raise "'#{cmd}' failed" unless system(cmd)
-        puts(cmd = "make || true 2>&1")
+        puts(cmd = "make 2>&1")
         raise "'#{cmd}' failed" unless system(cmd)
-        puts(cmd = "make install || true 2>&1")
+        puts(cmd = "make install 2>&1")
         raise "'#{cmd}' failed" unless system(cmd)
       end
       system("rm -rf #{BUNDLE_PATH}")
     end
-  end
-  
-  # Absolutely prevent the linker from picking up any other libmemcached
-  Dir.chdir("#{HERE}/lib") do
-    system("cp -f libmemcached.a libmemcached_gem.a") 
-    system("cp -f libmemcached.la libmemcached_gem.la") 
-  end
-  $LIBS << " -lmemcached_gem"
+  end  
 end
 
-if DARWIN
+case TARGET
+when :darwin
   $CFLAGS.gsub! /-arch \S+/, ''
   $CFLAGS << " -arch i386"
   $LDFLAGS.gsub! /-arch \S+/, ''
